+++ ElvUI_DoctorioUI/Core/Compartment.lua
@@ -1,19 +1,7 @@
-local DoctorioUI, E, L, V, P, G = unpack((select(2, ...)))
+local _, E = unpack((select(2, ...)))
 
 -- OnClick
 function DoctorioUI_OnAddonCompartmentClick()
 	E:ToggleOptions()
 	E.Libs.AceConfigDialog:SelectGroup('ElvUI', 'DoctorioUI')
 end
-
---[[
-	function DoctorioUI_OnAddonCompartmentEnter()
-		Private:Print('OnEnter Test Print')
-	end
-]]
-
---[[
-	function DoctorioUI_OnAddonCompartmentLeave()
-		Private:Print('OnLeave Test Print')
-	end
-]]