+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### Version 2.00 [15-02-2026]
+Added "Better Blizzard" Texture (Credits: Unhalted)  
+Added ElvUI Nameplates profile  
+Added ruRU localization (Credits: ZamestoTV)  
+Removed Method Raid Tools profile  
+Removed OmniCD profile  
+Removed WeakAuras imports  
+Reworked ElvUI import for Midnight  
+Updated Console Variables for Midnight  
+Updated Details profile  
+Updated installer steps  
+___
 ### Version 1.37 [26-09-2025]
 Re-run of the ElvUI profile install is highly recommended  
 Updated filters for ElvUI 14.00+